// src/app/model/emp.model.ts
export class Employee {
  id?: number;
  firstName?: string;
  dept?: string;
  degn?: string;
  email?: string;
  password?: string;
  options?: string;
}
export class EmployeeDetails {
  id?: number;
  firstName?: string;
  lastName?: string;
  mobile?: string;
  email?: string;
  address?: string;
  password: String;
    username: any;
    userType: String;
}
export class AttendanceModel{
  id?: number;
  name?: string;
  dept?: string;
  degn?: string;
  attendance?: string;
  leavetype?: string;
}

export class EditleaveModel{
  empiD? : number;
  empname? :string;
  empCode? : number;
  empcardNo?: number;
  dept? : string;
  fromdate? : string;
  todate : string;
  leavetype : string;
  status : string;
}


  